requirements.txt
argparse
matplotlib
python-shogi
setuptools
tensorboard
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchaudio==0.10.0+cu113
